fix(oauth): Fix UnsatifiedDependencyException (#393)
It looks like the ExternalSslAwareEntryPoint component is being created even though the outer class has a ConditionalOnExpression which returns false. I guess as it is a different class found by spring on startup and it doesn't consider it related to the outer class.

So i've simply duplicated the ConditionalOnExpression from the outer class to ExternalSslAwareEntryPoint.