fix(credentials): Remove hystrix from `CredentialsService` (#518)
There is no longer a remote call in the direct call path of
`CredentialsService.getAccounts()`, thus no need for hystrix.

The `CredentialsService` was originally getting called for every x.509
authentication request. The default hystrix configuration had placed
an artificial limit on the number of concurrent api requests that could
be served by `gate`.